# Introduce "pre-compiled" terms for buy gas, coinbase