--- a/test/Chainweb/Test/CoinContract.hs
+++ b/test/Chainweb/Test/CoinContract.hs
@@ -31,17 +31,13 @@ import Pact.Types.Runtime
 
 
 -- internal chainweb modules
-
-import Chainweb.Miner.Pact
 import Chainweb.Pact.TransactionExec
 
 
 tests :: TestTree
 tests = testGroup "Chainweb.Test.CoinContract"
   [ testGroup "Pact Command Parsing"
-    [ testCase "Buy Gas" buyGas'
-    , testCase "Coinbase" coinbase'
-    , testCase "Build Exec with Data" buildExecWithData
+    [ testCase "Build Exec with Data" buildExecWithData
     , testCase "Build Exec without Data" buildExecWithoutData
     ]
   , testGroup "Pact Code Unit Tests"
@@ -50,12 +46,6 @@ tests = testGroup "Chainweb.Test.CoinContract"
     , testCase "Payer Repl Tests" (ccReplTests "pact/gas-payer/gas-payer-v1.repl")
     ]
   ]
-
-buyGas' :: Assertion
-buyGas' = void $ mkBuyGasCmd minerId0 minerKeys0 sender0 1.0
-
-coinbase' :: Assertion
-coinbase' = void $ mkCoinbaseCmd minerId0 minerKeys0 1.0
 
 buildExecWithData :: Assertion
 buildExecWithData = void $ buildExecParsedCode
@@ -76,18 +66,3 @@ ccReplTests ccFile = do
         traverse_ (uncurry failCC) $ trFailure tr
 
     failCC i e = assertFailure $ renderInfo (_faInfo i) <> ": " <> unpack e
-
-------------------------------------------------------------------------------
--- Test Data
-------------------------------------------------------------------------------
-
-sender0 :: Text
-sender0 = "sender"
-
-minerKeys0 :: MinerKeys
-minerKeys0 = MinerKeys $ mkKeySet
-  ["f880a433d6e2a13a32b6169030f56245efdd8c1b8a5027e9ce98a88e886bef27"]
-  "default"
-
-minerId0 :: MinerId
-minerId0 = MinerId "default miner"