Gets a ~1ms savings per tx, and ~3ms per block

Before:
```
benchmarking PactService/block-new-valid[1]
time                 264.2 ms   (262.9 ms .. 265.3 ms)
                     
benchmarking PactService/block-new-valid[10]
time                 309.0 ms   (296.8 ms .. 325.9 ms)
                     
benchmarking PactService/block-new-valid[50]
time                 506.9 ms   (497.7 ms .. 514.7 ms)
                     
benchmarking PactService/block-new-valid[100]
time                 742.6 ms   (705.6 ms .. 778.2 ms)
                     
benchmarking PactService/block-new[1]
time                 134.1 ms   (132.9 ms .. 135.8 ms)
                     
benchmarking PactService/block-new[10]
time                 158.8 ms   (156.8 ms .. 163.9 ms)
                     
benchmarking PactService/block-new[50]
time                 247.4 ms   (242.8 ms .. 251.0 ms)
                     
benchmarking PactService/block-new[100]
time                 370.4 ms   (359.7 ms .. 388.2 ms)
```

After:
```
benchmarking PactService/block-new-valid[1]
time                 248.6 ms   (233.6 ms .. 269.9 ms)
                      
benchmarking PactService/block-new-valid[10]
time                 288.6 ms   (267.9 ms .. 308.2 ms)
                      
benchmarking PactService/block-new-valid[50]
time                 483.7 ms   (455.1 ms .. 550.2 ms)
                      
benchmarking PactService/block-new-valid[100]
time                 678.7 ms   (605.1 ms .. 763.6 ms)
                      
benchmarking PactService/block-new[1]
time                 117.7 ms   (113.3 ms .. 125.8 ms)
                      
benchmarking PactService/block-new[10]
time                 135.4 ms   (133.3 ms .. 139.7 ms)
                      
benchmarking PactService/block-new[50]
time                 246.8 ms   (230.6 ms .. 265.7 ms)
                      
benchmarking PactService/block-new[100]
time                 378.9 ms   (352.1 ms .. 438.9 ms)
```