@slpopejoy 2019-11-30
solid!
```
benchmarking PactService/block-new-valid[1]
time                 233.6 ms   (228.5 ms .. 237.9 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 229.5 ms   (225.4 ms .. 232.1 ms)
std dev              4.239 ms   (3.011 ms .. 5.524 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking PactService/block-new-valid[10]
time                 275.2 ms   (266.1 ms .. 282.8 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 270.0 ms   (268.0 ms .. 272.6 ms)
std dev              3.223 ms   (2.591 ms .. 3.849 ms)
variance introduced by outliers: 16% (moderately inflated)

benchmarking PactService/block-new-valid[50]
time                 451.7 ms   (441.4 ms .. 464.2 ms)
                     1.000 R²   (1.000 R² .. NaN R²)
mean                 456.9 ms   (454.3 ms .. 459.5 ms)
std dev              3.248 ms   (209.6 μs .. 3.987 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking PactService/block-new-valid[100]
time                 698.3 ms   (686.6 ms .. 705.7 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 695.1 ms   (691.7 ms .. 697.3 ms)
std dev              3.410 ms   (1.602 ms .. 4.763 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking PactService/block-new[1]
time                 113.6 ms   (109.1 ms .. 117.2 ms)
                     0.998 R²   (0.997 R² .. 1.000 R²)
mean                 116.6 ms   (114.5 ms .. 120.3 ms)
std dev              4.599 ms   (1.031 ms .. 6.932 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmarking PactService/block-new[10]
time                 135.4 ms   (133.7 ms .. 137.4 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 134.9 ms   (134.5 ms .. 135.6 ms)
std dev              840.5 μs   (594.4 μs .. 1.173 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmarking PactService/block-new[50]
time                 231.4 ms   (228.7 ms .. 233.2 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 232.0 ms   (231.3 ms .. 233.2 ms)
std dev              1.297 ms   (634.6 μs .. 1.842 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking PactService/block-new[100]
time                 353.6 ms   (350.2 ms .. 359.8 ms)
                     1.000 R²   (NaN R² .. 1.000 R²)
mean                 352.6 ms   (351.7 ms .. 353.5 ms)
std dev              1.042 ms   (886.4 μs .. 1.179 ms)
variance introduced by outliers: 19% (moderately inflated)

```