@slpopejoy 2019-11-30
Solid improvement!
```
benchmarking PactService/block-new-valid[1]
time                 233.6 ms   (228.5 ms .. 237.9 ms)

benchmarking PactService/block-new-valid[10]
time                 275.2 ms   (266.1 ms .. 282.8 ms)

benchmarking PactService/block-new-valid[50]
time                 451.7 ms   (441.4 ms .. 464.2 ms)

benchmarking PactService/block-new-valid[100]
time                 698.3 ms   (686.6 ms .. 705.7 ms)

benchmarking PactService/block-new[1]
time                 113.6 ms   (109.1 ms .. 117.2 ms)

benchmarking PactService/block-new[10]
time                 135.4 ms   (133.7 ms .. 137.4 ms)

benchmarking PactService/block-new[50]
time                 231.4 ms   (228.7 ms .. 233.2 ms)

benchmarking PactService/block-new[100]
time                 353.6 ms   (350.2 ms .. 359.8 ms)

```